backend/requirements_railway.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
python-dotenv==1.0.0
psycopg2-binary==2.9.9
pydantic==2.5.0
pydantic-settings==2.1.0
python-multipart==0.0.6
requests==2.31.0
pillow==10.1.0
aiofiles==23.2.1  # For file uploads
